+++ pnet-data-api-java/src/main/java/pnet/data/api/activity/ActivityDataFind.java
@@ -9,6 +9,7 @@ import pnet.data.api.util.RestrictBrand;
 import pnet.data.api.util.RestrictCompanyType;
 import pnet.data.api.util.RestrictContractType;
 import pnet.data.api.util.RestrictMatchcode;
+import pnet.data.api.util.RestrictScroll;
 import pnet.data.api.util.RestrictTenant;
 import pnet.data.api.util.RestrictUpdatedAfter;
 
@@ -20,7 +21,7 @@ import pnet.data.api.util.RestrictUpdatedAfter;
 public class ActivityDataFind extends AbstractFind<ActivityItemDTO, ActivityDataFind>
     implements RestrictMatchcode<ActivityDataFind>, RestrictTenant<ActivityDataFind>, RestrictBrand<ActivityDataFind>,
     RestrictCompanyType<ActivityDataFind>, RestrictContractType<ActivityDataFind>,
-    RestrictUpdatedAfter<ActivityDataFind>
+    RestrictUpdatedAfter<ActivityDataFind>, RestrictScroll<ActivityDataFind>
 {
 
     public ActivityDataFind(FindFunction<ActivityItemDTO> findFunction, List<Pair<String, Object>> restricts)