+++ pnet-data-api-java/src/main/java/pnet/data/api/company/CompanyDataFind.java
@@ -14,6 +14,7 @@ import pnet.data.api.util.RestrictHeadquarter;
 import pnet.data.api.util.RestrictId;
 import pnet.data.api.util.RestrictLocation;
 import pnet.data.api.util.RestrictPostalCode;
+import pnet.data.api.util.RestrictScroll;
 import pnet.data.api.util.RestrictTenant;
 import pnet.data.api.util.RestrictType;
 import pnet.data.api.util.RestrictUpdatedAfter;
@@ -27,7 +28,8 @@ public class CompanyDataFind extends AbstractFind<CompanyItemDTO, CompanyDataFin
     implements RestrictId<CompanyDataFind>, RestrictTenant<CompanyDataFind>, RestrictBrand<CompanyDataFind>,
     RestrictCompanyNumber<CompanyDataFind>, RestrictPostalCode<CompanyDataFind>, RestrictCountryCode<CompanyDataFind>,
     RestrictType<CompanyDataFind>, RestrictContractType<CompanyDataFind>, RestrictLocation<CompanyDataFind>,
-    RestrictExternalBrand<CompanyDataFind>, RestrictHeadquarter<CompanyDataFind>, RestrictUpdatedAfter<CompanyDataFind>
+    RestrictExternalBrand<CompanyDataFind>, RestrictHeadquarter<CompanyDataFind>, RestrictUpdatedAfter<CompanyDataFind>,
+    RestrictScroll<CompanyDataFind>
 {
 
     public CompanyDataFind(FindFunction<CompanyItemDTO> findFunction, List<Pair<String, Object>> restricts)