+++ pnet-data-api-java/src/main/java/pnet/data/api/util/RestrictScroll.java
@@ -1,0 +1,17 @@
+package pnet.data.api.util;
+
+/**
+ * Enables scrolling
+ *
+ * @author ham
+ * @param <SELF> the type of the restrict for chaining
+ */
+public interface RestrictScroll<SELF extends Restrict<SELF>> extends Restrict<SELF>
+{
+
+    default SELF scroll()
+    {
+        return restrict("scroll", true);
+    }
+
+}